tools/adk_docs_crawler/requirements.txt
# ADK Docs Crawler Dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# GCP
google-cloud-storage>=2.10.0

# Configuration
PyYAML>=6.0

# Utilities
urllib3>=2.0.0
